Refactor codebase to reorganize project structure
Moved business logic into dedicated internal packages, centralized configuration and logging setup in dedicated modules. Updated module path.
