+++ stock_scheduler_service/go.mod
@@ -1,15 +1,14 @@
-module stock_scheduler_service
+module github.com/igorbia/stock_scheduler_service
 
 go 1.24
 
 require (
-	github.com/jackc/pgpassfile v1.0.0 // indirect
-	github.com/jackc/pgservicefile v0.0.0-20240606120523-5a60cdf6a761 // indirect
-	github.com/jackc/pgx/v5 v5.7.4 // indirect
-	github.com/joho/godotenv v1.5.1 // indirect
-	github.com/lib/pq v1.10.9 // indirect
-	github.com/sirupsen/logrus v1.9.3 // indirect
-	golang.org/x/crypto v0.35.0 // indirect
+	github.com/joho/godotenv v1.5.1
+	github.com/lib/pq v1.10.9
+	github.com/sirupsen/logrus v1.9.3
+)
+
+require (
+	github.com/stretchr/testify v1.8.1 // indirect
 	golang.org/x/sys v0.30.0 // indirect
-	golang.org/x/text v0.22.0 // indirect
 )
