+++ stock_scheduler_service/log/logger.go
@@ -1,0 +1,13 @@
+package log
+
+import (
+	"os"
+
+	log "github.com/sirupsen/logrus"
+)
+
+func InitLogger() {
+	log.SetFormatter(&log.JSONFormatter{})
+	log.SetOutput(os.Stdout)
+	log.SetLevel(log.InfoLevel)
+}
